🩹 update require auth logic
