+++ src/pages/knowledge/[id]/diff/[diff].tsx
@@ -13,10 +13,7 @@ import dayjs from "dayjs"
 import relativeTime from "dayjs/plugin/relativeTime"
 import { GetServerSideProps } from "next"
 import Link from "next/link"
-import { useRouter } from "next/router"
 import { getServerSession } from "next-auth"
-import { useSession } from "next-auth/react"
-import { useEffect } from "react"
 import { remark } from "remark"
 import remarkBreaks from "remark-breaks"
 import remarkGfm from "remark-gfm"
@@ -33,15 +30,6 @@ type Props = {
 const Page: NextPageWithLayout<Props> = (props) => {
   const { id: diffid, title, createdAt } = props.diff
   const { id } = props.knowledge
-
-  const { data: session } = useSession()
-  const router = useRouter()
-
-  useEffect(() => {
-    if (!session && typeof session != "undefined") {
-      router.push(`/`)
-    }
-  }, [session, router])
 
   return (
     <>
@@ -71,10 +59,13 @@ const Page: NextPageWithLayout<Props> = (props) => {
 export const getServerSideProps: GetServerSideProps = async ({ params, req, res }) => {
   const session = await getServerSession(req, res, authOptions)
 
-  if (!session) {
-    res.statusCode = 403
-    return { props: { knowledge: [] } }
-  }
+  if (!session)
+    return {
+      redirect: {
+        destination: "/",
+        permanent: false,
+      },
+    }
 
   const knowledgedata = await prisma.knowledge.findFirst({
     include: {
