+++ src/pages/knowledge/[id]/diff/index.tsx
@@ -12,10 +12,7 @@ import dayjs from "dayjs"
 import relativeTime from "dayjs/plugin/relativeTime"
 import { GetServerSideProps } from "next"
 import Link from "next/link"
-import { useRouter } from "next/router"
 import { getServerSession } from "next-auth"
-import { useSession } from "next-auth/react"
-import { useEffect } from "react"
 import { GrHistory } from "react-icons/gr"
 import { IoTimeOutline } from "react-icons/io5"
 
@@ -24,16 +21,6 @@ dayjs.locale("ja")
 
 const Page: NextPageWithLayout<KnowledgeProps> = (props) => {
   const { id, title, diff } = props
-
-  const { data: session } = useSession()
-
-  const router = useRouter()
-
-  useEffect(() => {
-    if (!session && typeof session != "undefined") {
-      router.push(`/`)
-    }
-  }, [session, router])
 
   return (
     <>
@@ -126,10 +113,13 @@ Page.getLayout = (page) => <Layout>{page}</Layout>
 export const getServerSideProps: GetServerSideProps = async ({ params, req, res }) => {
   const session = await getServerSession(req, res, authOptions)
 
-  if (!session) {
-    res.statusCode = 403
-    return { props: { knowledge: [] } }
-  }
+  if (!session)
+    return {
+      redirect: {
+        destination: "/",
+        permanent: false,
+      },
+    }
 
   const data = await prisma.knowledge.findFirst({
     include: {
