+++ src/pages/tags/index.tsx
@@ -55,9 +55,14 @@ const Page: NextPageWithLayout<Props> = (props) => {
 
 export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
   const session = await getServerSession(req, res, authOptions)
-  if (!session) {
-    return { props: { tags: [] } }
-  }
+
+  if (!session)
+    return {
+      redirect: {
+        destination: "/",
+        permanent: false,
+      },
+    }
 
   const tags = await prisma.tag.findMany({
     orderBy: [
